pyproject.toml
[project]
name = "hybrid-rag-structured"
version = "0.1.0"
description = "Hybrid RAG: semantic + structured retrieval over Postgres"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "sqlalchemy>=2.0.0",
    "tantivy>=0.22.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
